arcella::config

Function add_no_redef_warning

fn add_no_redef_warning(
    key: String,
    config_idx: usize,
    config_files: &IndexSet<PathBuf>,
    source_file_idx: usize,
    warnings: &mut Vec<ConfigLoadWarning>,
)

Adds a warning that a value was ignored because the corresponding key was not marked with #redef in the main configuration layer.

Arguments

key – The configuration key that was ignored.

config_idx – The index of the config (in the configs vector) that lacked the #redef permission.

config_files – The global set of loaded config files (used to resolve file indices to paths).

source_file_idx – The index of the file that attempted to set the value (used for diagnostics).

warnings – Mutable vector to which the warning will be appended.